[bug] retry the sync insertion loop in case of InvalidGenerationError
If the generation changes while a sync is happening the sync fails with
InvalidGenerationError. This can happen if there is an insertion or
another operation before the sync finishes. The sync was going to an
error loop.

The proposed solution is to retry the sync loop in case of failure. As
collateral effect if the insertion loop has not process all the docs we
retry as well.

Resolves: #6757